Add files via upload
Description: This C program demonstrates the implementation of Bubble Sort algorithm. In this program, an array of integers is sorted in ascending order using the Bubble Sort algorithm.

Algorithm:

Declare a function named Bubble that accepts an integer array and the size of the array as arguments. Declare a variable named passes and set its value to n - 1 where n is the size of the array. Declare variables named comparisons, i, j, and temp. Start a loop that will run for i from 0 to passes-1. Set the value of comparisons to n - 1 - i. Start a loop that will run for j from 0 to comparisons-1. Check if a[j] is greater than a[j+1]. If true, swap the values of a[j] and a[j+1]. End both loops. Declare a function named main. Declare an integer array named a and initialize it with some values. Declare an integer variable named n and set its value to 5. Call the Bubble function and pass the array a and value of n as arguments. Start a loop that will run for i from 0 to n-1. Print the value of each element of the array a. End the loop. End the main function.

Output: 20 41 54 119 123